remove compat filesystem after g++ update add better diagnostics to formatter succinct lints from windows macros tweak c++ bulk test windows types use ULONG over DWORD DWORD64\ULONG64 -> uint64_t add move semantics to avoid exception warnings use smaller enum types remove redundant code #pragma once

+++ Coverage/Line.h
@@ -2,7 +2,7 @@
 
 #include <GLib/Eval/Evaluator.h>
 
-enum class LineCover : int;
+enum class LineCover : uint8_t;
 
 struct Line
 {
