+++ Coverage/LineCover.h
@@ -3,7 +3,7 @@
 #include <array>
 #include <ostream>
 
-enum class LineCover : int
+enum class LineCover : uint8_t
 {
 	None,
 	Covered,
@@ -13,5 +13,5 @@ enum class LineCover : int
 inline std::ostream & operator<<(std::ostream & s, const LineCover & cov)
 {
 	constexpr auto a = std::array {std::string_view {}, std::string_view {"cov"}, std::string_view {"ncov"}};
-	return s << a.at(static_cast<int>(cov));
+	return s << a.at(static_cast<uint8_t>(cov));
 }
