+++ GLib/Fwd.h
@@ -1,6 +1,8 @@
 #pragma once
 
+#include <cstdint>
+
 namespace GLib::Flog
 {
-	enum class Level : unsigned;
+	enum class Level : uint8_t;
 }
